requirements.txt
certifi==2020.11.8
chardet==3.0.4
DateTime==4.3
idna==2.10
mysql-connector-python==8.0.22
oauthlib==3.1.0
protobuf==3.14.0
PySocks==1.7.1
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
tweepy==3.9.0
urllib3==1.26.2
zope.interface==5.2.0
